feat: #76 leverage conftest.py
