+++ tests/models/kepler_problem/test_dynamics.py
@@ -16,9 +16,6 @@ from hamilflow.models.kepler_problem.dynamics import (
     _tau_of_e_plus_u_elliptic,
     _tau_of_u_exact_elliptic,
     _tau_of_u_exact_hyperbolic,
-    tau_of_u_elliptic,
-    tau_of_u_hyperbolic,
-    tau_of_u_parabolic,
     tau_of_u_prime,
 )
 
@@ -29,38 +26,8 @@ if TYPE_CHECKING:
 
 _EPS_SQRT = 1e-16
 
-_EPS_ECC = 1e-5
 
-
-# 5 / 7, 12 / 11, 257 / 13 makes u_of_tau from newton fail
-@pytest.fixture(params=[0.0, 0.1, 0.3, 0.7, 0.9, 1.0, 1.1, 2.0, 11.0, 101.0])
-def ecc(request: pytest.FixtureRequest) -> float:
-    return request.param
-
-
-@pytest.fixture(params=[-0.9, False, 0.9])
-def u_s(request: pytest.FixtureRequest, ecc: float) -> "npt.ArrayLike":
-    # There are dividends sqrt(e**2 - u**2) and (u + 1),
-    # hence u cannot be too close to +e / -e / -1
-    f = 1 - _EPS_ECC
-    ecc_f = ecc * f
-    return max(-f, request.param * ecc_f) or np.linspace(max(-f, -ecc_f), ecc_f, 7)
-
-
-@pytest.fixture()
-def tau_of_u(ecc: float) -> "Callable[[float, npt.ArrayLike], npt.ArrayLike]":
-    if ecc == 0:
-        pytest.skip(f"Circular case")
-    elif 0 < ecc < 1:
-        return tau_of_u_elliptic
-    elif ecc == 1:
-        return tau_of_u_parabolic
-    elif ecc > 1:
-        return tau_of_u_hyperbolic
-    else:
-        raise ValueError(f"Expect ecc >= 0, got {ecc}")
-
-
+@pytest.mark.usefixtures("ecc", "u_s", "tau_of_u")
 class TestTauOfU:
     def test_const(
         self, ecc: float, tau_of_u: "Callable[[float, npt.ArrayLike], npt.ArrayLike]"
@@ -108,13 +75,14 @@ class TestTauOfU:
             raise ValueError(f"Expect ecc >= 0, got {ecc}")
 
     @pytest.mark.parametrize("epsilon", [1e-7])
-    @pytest.mark.skipif("ecc == 1.0")
     def test_expansion(
         self,
         ecc: float,
         epsilon: float,
         exact_and_approx_tau_s: "tuple[Callable[[float, npt.NDArray[np.float64]], npt.NDArray[np.float64]], Callable[[float, npt.NDArray[np.float64]], npt.NDArray[np.float64]], Callable[[float, npt.NDArray[np.float64]], npt.NDArray[np.float64]]]",
     ) -> None:
+        if ecc == 0.0 or ecc == 1.0:
+            pytest.skip(f"Test applies to ecc > 0, ecc != 1, got {ecc}")
         factor = 1 - epsilon
         f, g_s = exact_and_approx_tau_s[0], exact_and_approx_tau_s[1:]
         for u, g in zip((max(-1, -ecc) * factor, ecc * factor), g_s):
