+++ tests/models/kepler_problem/test_model.py
@@ -6,6 +6,8 @@ import pytest
 
 from hamilflow.models.kepler_problem import Kepler2DIoM, Kepler2DSystem
 
+from .conftest import *
+
 
 @pytest.fixture(params=[(1.0, 1.0), (1.0, 2.0), (2.0, 1.0), (2.0, 2.0)])
 def system_kwargs(request: pytest.FixtureRequest) -> dict[str, float]:
@@ -14,12 +16,6 @@ def system_kwargs(request: pytest.FixtureRequest) -> dict[str, float]:
 
 @pytest.fixture(params=[1.0, 2.0])
 def parameter(request: pytest.FixtureRequest) -> float:
-    return request.param
-
-
-# 5 / 7, 12 / 11, 257 / 13 makes u_of_tau from newton fail
-@pytest.fixture(params=[0.0, 0.1, 0.3, 0.7, 0.9, 1.0, 1.1, 2.0, 11.0, 101.0])
-def ecc(request: pytest.FixtureRequest) -> float:
     return request.param
